collapse prod/dev configureStore
